f64

+++ src/data_generators.jl
@@ -17,21 +17,21 @@ julia> x,y,clusters = generate_gaussian_data(10000,2,6,100.0)
 ```
 """
 function generate_gaussian_data(N::Int64, D::Int64, K::Int64,MixtureVar::Number)
-	x = randn(Float32,D,N)
+	x = randn(Float64,D,N)
 	tpi = rand(Dirichlet(ones(K)))
 	tzn = rand(Multinomial(N,tpi))
-	tz = zeros(Float32,N)
+	tz = zeros(Float64,N)
 
-	tmean = zeros(Float32,D,K)
-	tcov = zeros(Float32,D,D,K)
+	tmean = zeros(Float64,D,K)
+	tcov = zeros(Float64,D,D,K)
 
 	ind = 1
 	println(tzn)
 	for i=1:length(tzn)
 		indices = ind:ind+tzn[i]-1
 		tz[indices] .= i
-		tmean[:,i] .= rand(MvNormal(zeros(Float32,D), MixtureVar*Matrix{Float32}(I, D, D)))
-		tcov[:,:,i] .= rand(InverseWishart(D+2, Matrix{Float32}(I, D, D)))
+		tmean[:,i] .= rand(MvNormal(zeros(Float64,D), MixtureVar*Matrix{Float64}(I, D, D)))
+		tcov[:,:,i] .= rand(InverseWishart(D+2, Matrix{Float64}(I, D, D)))
 		d = MvNormal(tmean[:,i], tcov[:,:,i])
 		for j=indices
 			x[:,j] = rand(d)
@@ -58,7 +58,7 @@ julia> generate_mnmm_data(10000, 10, 5, 100)
 """
 function generate_mnmm_data(N::Int64, D::Int64, K::Int64, trials::Int64)
 	clusters = zeros(Float64,D,K)
-	x = zeros(Float32,D,N)
+	x = zeros(Float64,D,N)
 	labels = rand(1:K,(N,))
 	for i=1:K
 		alphas = rand(1:20,(D,))
